Support cell-level error handling
In the future we might want to make this optional, but for now let's not
consider invalid data in a cell a cause for throwing a 500. This is
because it seems to happen often enough, and it's really nice to give
feedback on exactly which cell(s) had bad data rather than show a
generic load error.