Fix not parsing some empty cells
If there are empty cells of 'Generic' type, excel doesn't write them to
the xml. This has the effect of shifting cell data left for that row
after that cell if unhandled, as was the case previously.

Rather than iterating over the xml data, this commit counts cells
explicitly by cell name (ex. 'A2') until the last cell, recording nil
when there is no xml data for a cell.