Don't require sheets to specify dimensions
Aka support reading simple_xlsx_writer docs.

Excel writes a 'dimension' node to worksheets which we use to get the
last column from, but either it's not strictly required or
simple_xlsx_writer ignores it.

Since we can usually assume the first row should be the same length as
other rows, this uses the last cell of the first row if dimension is
unspecified. Not an ideal assumption, but good enough for this already
edge case.