Improve empty row handling
Previously, empty rows would be ommitted.

This adds an empty row of the correct size.

Also, refactors to something that should be a speed improvement,
although it doesn't seem to be actually. Should be easier to improve
speed though, since we now only do one call to nokogiri, which at least
reduces the things we want to check.