Be more flexible in handling custom formatting
Previously, we would not accept a custom style with a numFmtId < 164.
This seems to be correct, but I suppose it depends on how you interpret
the spec (ISO 29500). In the wild I found a spreadsheet from an
offshore team whose software was generating cells with 59 as a custom
formatted date... oh well, maybe that's OK since it's not explicitly
defined in the "All Languages" predefined styles (it *is* defined in
the Thai localization).

Let's just be flexible, and say that if it's not explicitly defined in
the "All Languages" predefined styles, we will recognize it as a custom
format so long as it's been defined as such.