Raise an error when date styles are not numerical
A previous commit broke an edge case where we want to raise when a
date-formatted cell is not numerical, ex. the value "hello." This
probably doesn't happen in practice much, but it's a case we used to
handle, so let's keep handling it.

This fixes the issue, as well as tests expecting this behavior, and
adds some more focused unit tests to catch any similar regression
sooner and more explicitly.