Add ability to parse hyperlinks in cells
Previously, if someone put a hyperlink in a URL, we would parse the
"friendly name" and not the URL.

This changes it so that we prioritize parsing the URL, yet don't lose
the "friendly name" if we happen to want it.

PERFORMANCE NOTE: while this does not slow down previous functionality,
spreadsheets with many hyperlinks show to be at least an order of
magnitude slower (maybe two) as well as failing the linear performance
test (which doesn't actually care about overall speed). The performance
test, for now, just tests functionality sans-hyperlinks, but this is an
area that needs improvement.