Stream parsing overhaul
Previously, we couldn't load files over ~10k rows or the memory would
run away from us. This was because we were slurping all the workbooks
into Nokogiri nodes all at once, which was known to be inefficient, but
convenient.

This upgrades the shared strings and sheet parsing to use a SAX stream
parser, and then yield the rows without ever slurping any of this XML
into memory. Unfortunately we still need to hold the shared strings in
memory, but this is still a huge improvement.