Treat typeless cells as numbers by default
When a cell does not have a type, the Office Open XML standard states
that, by default, the type should be 'n' (number). This commit makes
sure that, when that happens, the cell value is cast to Integer or Float
by default - and, if not possible, left as a string.