Fix parsing "chunky" UTF-8 workbooks
For some reason I can't figure out in a reasonable timeframe,
SAX parsing some workbooks captures separate strings in the same cell
when we encounter UTF-8, although I can't get workbooks made in my
own version of excel to repro it. Our fix is just to keep building
the string in this case, although maybe there's a setting in Nokogiri
to make it not do this (looked, couldn't find it).

Loading the workbook test/chunky_utf8.xlsx repros the issue.

Closes #39 & #45