For hyperlinks, use visible cell value vs URL
Previously, we defaulted to extracting the URL value for hyperlinks and
making the visible text available in a "friendly_name" attribute. This
might make sense for actual hyperlinks, but it turns out email addresses
are often represented using the hyperlink function and our strategy is
surprising in that case, where it extracts the email including "mailto:"

This changes the hyperlink object (SimpleXlsxReader::Hyperlink) to by
default show what the user sees in the cell, and adds a #url method to
get the URL. This should be easy to check for when parsing spreadsheet
data if you want the URL, but won't surprise anyone by extracting
different data than what is seen visually in the spreadsheet.