25% performance improvement
In poking around at various ways to support namespaces, I tried the
namespace syntax for nokogiri, the "correct" way to do it. For some
reason this provides a 25% performance improvement. Measuring the memory
usage, there *might* be an ~8% increase in memory usage for big files. I
didn't make a statistically valid analysis of the difference, I just ran
each side of the diff through the big file test about 20 times back and
forth. Sometimes there was no difference, sometimes the performance
improvement used less memory also, but it seems like more often than not
it might use a bit more memory.

Every time, without fail, this change gave a clear 25% improvement, so a
20 second performance test would take 15s. Given that the memory usage
is not obviously much worse if at all, this is the "correct" way to
support namespaces, and it definitely gives us a huge performance win,
we're going to ship it.

Note that instead of passing in arrays of attribute strings, it passes
in a value object with #localname and #value. I could speculate on
the ways this might result in the improvements we're seeing, but I'm not
sure, it's just worth pointing out.